build: update scripts and config files
