--- .node-version
@@ -1,1 +1,0 @@
-16.15.0
