+++ package.json
@@ -20,41 +20,24 @@
     "./site"
   ],
   "scripts": {
-    "prebuild": "npm run clean",
-    "start": "chokidar \"src/**/*.ts\" -c \"npm run build\"",
+    "start": "echo 'TODO: chokidar \"src/**/*.ts\" -c \"npm run build\" || npm run test:watch'",
+    "prebuild": "npm run clean && npm run types-check",
     "build": "npm run build:cjs && npm run build:esm",
-    "build:cjs": "tsc -p tsconfig.build.json --module commonjs --outDir dist/cjs",
-    "build:esm": "tsc -p tsconfig.build.json --module esNext --outDir dist/esm",
+    "postbuild": "npm run copy:hooks && npm run update:readme",
+    "build:cjs": "tsc --module commonjs --outDir dist/cjs",
+    "build:esm": "tsc --module esNext --outDir dist/esm",
     "test": "jest",
+    "test:watch": "npm run test -- --watch --silent",
+    "test:coverage": "ts-node ./scripts/coverage.ts",
     "clean": "rimraf -rf ./dist",
-    "types-check": "tsc -p tsconfig.json",
-    "legacy-clean": "lerna run clean && lerna clean -y",
-    "legacy-bootstrap": "npm ci && lerna exec npm ci && npm run copy:hooks && npm run build:lib",
-    "legacy-types-check": "lerna run types-check",
-    "legacy-lint": "npm-run-all --continue-on-error -p lint:*",
-    "legacy-lint:code": "eslint '**/*.{js,jsx,ts,tsx}'",
-    "legacy-lint:other": "prettier --check \"**/*.{json,md,mdx,css,scss,yaml,yml}\"",
-    "legacy-format": "run-p format:*",
-    "legacy-format:code": "npm run lint:code -- --fix",
-    "legacy-format:other": "npm run lint:other -- --write",
-    "legacy-pretest": "npm run types-check && npm run format",
-    "legacy-test": "jest",
-    "legacy-test:ssr": "jest --config ./jest.config.node.ts",
-    "legacy-test:watch": "npm run test -- --watch --silent",
-    "legacy-test:coverage": "ts-node ./scripts/coverage.ts",
-    "legacy-prebuilt": "npm run format && npm run copy:hooks",
-    "legacy-build": "lerna run build --stream",
-    "legacy-build:site": "lerna run --scope site build",
-    "legacy-build:lib": "lerna run --scope usehooks-ts build && npm run copy:hooks",
-    "legacy-predeploy": "npm run build",
-    "legacy-deploy": "netlify deploy --dir=site/public --prod",
-    "legacy-dev:site": "npm run build:lib && cd site && npm run start",
-    "legacy-dev:lib": "npm run test:watch",
-    "legacy-serve:site": "cd site && npm run serve",
-    "legacy-publish": "npm run build:lib && npm run test && lerna publish --no-git-reset",
-    "legacy-update:readme": "ts-node ./scripts/updateReadme.ts",
-    "legacy-copy:hooks": "rimraf -rf ./site/generated && ts-node ./scripts/copyHooks.ts",
-    "legacy-plop": "plop && npm run update:readme && npm run format"
+    "format": "prettier --check \"**/*.{json,md,mdx,css,scss,yaml,yml}\"",
+    "lint": "eslint '**/*.{js,jsx,ts,tsx}'",
+    "types-check": "tsc --noEmit",
+    "publish": "echo 'TODO: Publish npm package on npm.com'",
+    "update:readme": "ts-node ./scripts/updateReadme.ts",
+    "copy:hooks": "rimraf -rf ./site/generated && ts-node ./scripts/copyHooks.ts",
+    "plop": "plop && npm run update:readme && npm run format",
+    "deploy-site": "echo 'TODO: deploy: netlify deploy --dir=site/public --prod'"
   },
   "devDependencies": {
     "@testing-library/react": "^12.1.2",
@@ -92,6 +75,10 @@
   },
   "peerDependencies": {
     "react": "^16.9.0 || ^17.0.0 || ^18.0.0"
+  },
+  "engines": {
+    "node": ">=16.15.0",
+    "npm": ">=8"
   },
   "repository": {
     "type": "git",
