--- tsconfig.base.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "sourceMap": true,
-    "declaration": true,
-    "declarationMap": true,
-    "noEmitOnError": true,
-    "skipLibCheck": true,
-    "removeComments": true,
-    "esModuleInterop": true,
-    "jsx": "react",
-    "strict": true,
-    "isolatedModules": true,
-    "lib": ["ESNEXT", "DOM"]
-  },
-
-  "exclude": ["node_modules", "dist"]
-}
