--- tsconfig.build.json
@@ -1,15 +1,0 @@
-{
-  "extends": "./tsconfig.base.json",
-
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist",
-    "target": "esnext",
-    "moduleResolution": "node",
-    // Needed for CommonJS modules
-    "allowSyntheticDefaultImports": true
-  },
-
-  "include": ["./src/**/*"],
-  "exclude": ["./src/**/*.test.ts"]
-}
